requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
configparser==3.7.3
Django==2.1.7
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
gunicorn==19.9.0
httplib2==0.12.1
idna==2.8
lxml==4.3.1
mysql-connector
mysqlclient==1.4.2.post1
Pillow==5.4.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
PyMySQL==0.9.3
PySocks==1.6.8
pytz==2018.9
PyYAML==5.1
requests==2.21.0
rsa==4.0
six==1.12.0
soupsieve==1.8
uritemplate==3.0.0
urllib3==1.24.1
